// src/app/core/api/model/category.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tenant } from './tenant';


export interface Category { 
    id?: number;
    name?: string;
    description?: string;
    tenant?: Tenant;
}